agent/requirements.txt
anthropic>=0.89.0
notion-client>=3.0.0
todoist-api-python>=4.0.0
PyGithub>=2.9.0
google-api-python-client>=2.193.0
google-auth-httplib2>=0.3.1
google-auth-oauthlib>=1.3.1
python-dotenv>=1.1.0
